# local/frontend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AIAnalysisResult": {
        "additionalProperties": false,
        "description": "Stable result envelope for validated AI or deterministic fallback.",
        "properties": {
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "fallback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeterministicAIFallback"
              },
              {
                "type": "null"
              }
            ]
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "result_type": {
            "$ref": "#/components/schemas/AIAnalysisResultType"
          },
          "validated_ai": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidatedAIAnalysis"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "result_type",
          "incident_id"
        ],
        "title": "AIAnalysisResult",
        "type": "object"
      },
      "AIAnalysisResultType": {
        "enum": [
          "VALIDATED_AI",
          "FALLBACK"
        ],
        "title": "AIAnalysisResultType",
        "type": "string"
      },
...
      "DeterministicAIFallback": {
        "additionalProperties": false,
        "description": "Typed non-model result assembled only from the EvidenceBundle.",
        "properties": {
          "evidence_bundle_status": {
            "$ref": "#/components/schemas/EvidenceBundleStatus"
          },
          "exact_history": {
            "items": {
              "$ref": "#/components/schemas/FallbackEvidenceFact"
            },
            "title": "Exact History",
            "type": "array"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "primary_evidence": {
            "items": {
              "$ref": "#/components/schemas/FallbackEvidenceFact"
            },
            "title": "Primary Evidence",
            "type": "array"
          },
          "reason": {
            "$ref": "#/components/schemas/AIAnalysisFallbackReason"
          },
          "reason_detail": {
            "title": "Reason Detail",
            "type": "string"
          },
          "semantic_history": {
            "$ref": "#/components/schemas/SemanticEnrichmentSummary"
          },
          "verification": {
            "items": {
              "$ref": "#/components/schemas/FallbackVerificationFact"
            },
            "title": "Verification",
            "type": "array"
          }
        },
        "required": [
          "reason",
          "reason_detail",
          "incident_id",
          "evidence_bundle_status",
          "primary_evidence",
          "exact_history",
          "semantic_history",
          "verification"
        ],
        "title": "DeterministicAIFallback",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentAuditAction": {
        "enum": [
          "INCIDENT_CREATED",
          "EVIDENCE_LINKED",
          "EVIDENCE_UNLINKED",
          "INCIDENT_SPLIT",
          "STATUS_CHANGED",
          "OWNER_CHANGED",
          "SEVERITY_CHANGED",
          "DUE_STATE_CHANGED",
          "DUE_TIME_CHANGED",
          "RECURRENCE_RECORDED",
          "HANDOVER_ACKNOWLEDGED"
        ],
        "title": "IncidentAuditAction",
        "type": "string"
      },
      "IncidentAuditOutput": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/IncidentAuditAction"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "id",
          "action",
          "occurred_at",
          "payload"
        ],
        "title": "IncidentAuditOutput",
        "type": "object"
      },
      "IncidentAuditResponse": {
        "additionalProperties": false,
        "properties": {
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/IncidentAuditOutput"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "audit_events"
        ],
        "title": "IncidentAuditResponse",
        "type": "object"
      },
      "IncidentCollectionResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "IncidentCollectionResponse",
        "type": "object"
      },
      "IncidentDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/IncidentAuditOutput"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due State"
          },
          "due_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "owner_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Ref"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "machine_id",
          "status",
          "owner_ref",
          "severity",
          "due_state",
          "due_time",
          "created_at",
          "updated_at",
          "audit_events"
        ],
        "title": "IncidentDetailResponse",
        "type": "object"
      },
...
      "IncidentEvidenceResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/IncidentEvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "evidence"
        ],
        "title": "IncidentEvidenceResponse",
        "type": "object"
      },
      "IncidentStatus": {
        "enum": [
          "OPEN",
          "VERIFYING",
          "VERIFIED",
          "RECURRED"
        ],
        "title": "IncidentStatus",
        "type": "string"
      },
      "IncidentSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due State"
          },
          "due_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "owner_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Ref"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "machine_id",
          "status",
          "owner_ref",
          "severity",
          "due_state",
          "due_time",
          "created_at",
          "updated_at"
        ],
        "title": "IncidentSummaryResponse",
        "type": "object"
      },
      "JsonValue": {},
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VerificationRuleType": {
        "enum": [
          "NO_EVENT"
        ],
        "title": "VerificationRuleType",
        "type": "string"
      },
      "VerificationRunResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "evidence_event_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Event Ids",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationRunResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "rule_identifier": {
            "title": "Rule Identifier",
            "type": "string"
          },
          "rule_name": {
            "title": "Rule Name",
            "type": "string"
          },
          "rule_type": {
            "$ref": "#/components/schemas/VerificationRuleType"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "verification_rule_id": {
            "format": "uuid",
            "title": "Verification Rule Id",
            "type": "string"
          },
          "window_ends_at": {
            "format": "date-time",
            "title": "Window Ends At",
            "type": "string"
          },
          "window_minutes": {
            "title": "Window Minutes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "incident_id",
          "verification_rule_id",
          "rule_identifier",
          "rule_name",
          "rule_type",
          "window_minutes",
          "result",
          "started_at",
          "window_ends_at",
          "completed_at",
          "evidence_event_ids"
        ],
        "title": "VerificationRunResponse",
        "type": "object"
      },
      "VerificationRunResult": {
        "enum": [
          "SUCCEEDED",
          "RECURRENCE_DETECTED"
        ],
        "title": "VerificationRunResult",
        "type": "string"
      },
      "VerificationRunsResponse": {
        "additionalProperties": false,
        "properties": {
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/VerificationRunResponse"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "incident_id",
          "runs"
        ],
        "title": "VerificationRunsResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MINE-TRACE Backend",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "machine_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "owner_ref",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Ref"
            }
          },
          {
            "in": "query",
            "name": "due_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/ai-analysis": {
      "post": {
        "description": "Run optional advisory analysis only after an explicit POST action.",
        "operationId": "analyze_incident_evidence_api_v1_incidents__incident_id__ai_analysis_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Incident Evidence",
        "tags": [
          "AI Insights"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/audit": {
      "get": {
        "operationId": "get_incident_audit_api_v1_incidents__incident_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentAuditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Audit",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/evidence": {
      "get": {
        "operationId": "get_incident_evidence_api_v1_incidents__incident_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Evidence",
        "tags": [
          "Incidents"
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}/verifications": {
      "get": {
        "operationId": "list_verifications_api_v1_incidents__incident_id__verifications_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationRunsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Verifications",
        "tags": [
          "Verification"
        ]
      }
    },
...
    "/api/v1/verifications/{run_id}": {
      "get": {
        "operationId": "get_verification_api_v1_verifications__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Verification",
        "tags": [
          "Verification"
        ]
      }
    },
...
  },
...
}